requirements.txt
fastapi
sqlalchemy
uvicorn
python-jose
passlib
bcrypt<4.0
python-dotenv
authlib
pydantic[email]
httpx
itsdangerous